python-backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.3
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.67
botocore==1.35.67
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==4.4.2
distro==1.9.0
filelock==3.16.1
filetype==1.2.0
fire==0.7.0
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.7
imageio==2.36.0
imageio-ffmpeg==0.5.1
imgaug==0.4.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
openai==1.54.4
openai-whisper @ git+https://github.com/openai/whisper.git@173ff7dd1d9fb1c4fddea0d41d704cfefeb8908c
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pillow==11.0.0
proglog==0.1.10
propcache==0.2.0
protobuf==3.20.2
psutil==6.1.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.49
s3transfer==0.10.4
safetensors==0.4.5
sahi==0.11.19
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.1.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.35
ultralytics-thop==2.0.12
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.17.2